+++ R/utils.R
@@ -29,14 +29,13 @@ confint.medoutcon <- function(object,
 
   # assume continuous outcome if more than two levels in outcome node
   if (length(unique(object$outcome)) > 2 ||
-    object$param %in% c("direct_effect", "indirect_effect")) {
+      stringr::str_detect(object$param, "direct")) {
     # NOTE: variance already scaled (i.e., Var(D)/n)
     se_eif <- sqrt(object$var)
 
     # compute the interval around the point estimate
     ci_theta <- ci_norm_bounds * se_eif + object$theta
-  } else if (length(unique(object$outcome)) == 2 &&
-    !(object$param %in% c("direct_effect", "indirect_effect"))) {
+  } else if (length(unique(object$outcome)) == 2) {
     # for binary outcomes, create CI on the logit scale and back-transform
     theta_ratio <- stats::qlogis(object$theta)
     grad_ratio_delta <- (1 / object$theta) + (1 / (1 - object$theta))